LPD-77757 Replacing at batch file level
For this case the data from the WidgetInstance was directly used to
create portlet preferences and fetch a portlet. The current replacement
mechanism cannot fix this as it is based on model listeners and it's
already too late.

With this change we make the replacements on the batch file and then
send it for processing as usual.
