+++ modules/apps/batch-engine/batch-engine-api/src/main/java/com/liferay/batch/engine/BatchEngineFileProcessor.java
@@ -1,0 +1,15 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.batch.engine;
+
+/**
+ * @author Alejandro Tardín
+ */
+public interface BatchEngineFileProcessor {
+
+	public String process(String content);
+
+}
