+++ modules/apps/object/object-api/bnd.bnd
@@ -11,6 +11,7 @@ Export-Package:\
 	com.liferay.object.constants,\
 	com.liferay.object.definition.groovy.script.use,\
 	com.liferay.object.definition.notification.term.util,\
+	com.liferay.object.definition.processor,\
 	com.liferay.object.definition.security.permission.resource,\
 	com.liferay.object.definition.security.permission.resource.util,\
 	com.liferay.object.definition.setting.builder,\
