+++ modules/apps/object/object-api/src/main/java/com/liferay/object/definition/processor/ObjectDefinitionClassNameProcessor.java
@@ -1,0 +1,17 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.object.definition.processor;
+
+/**
+ * @author Alejandro Tardín
+ */
+public interface ObjectDefinitionClassNameProcessor {
+
+	public long updateObjectDefinitionClassNameId(long classNameId);
+
+	public String updateObjectDefinitionReferences(String value);
+
+}
