+++ modules/apps/object/object-api/src/main/resources/com/liferay/object/definition/processor/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
